Fixed Spencer's Workaround
Spencer's workaround was good in theory, but somewhere along the line I
think he got lost in the code? The function was loading and unloading
the library every time anyway if the function was entered with an
unloaded library. I fixed it so now it checks to see if the library
needs to be loaded, and if it does, loads it. Also, the library has to
be loaded anew every time the motors get plugged in, so I added a
function unload_library that does just that.

I also generalised GoToPos and ReturnToPos so we can all use it by
assigning the path to a variable and calling the variable. I've
defaulted it back to Spencer after I was done.

Moving the dynamixels now is fast - so fast, actually, that I can brick
the motor by not giving it time to pause using codytest. The limiting
factor is loading the robotics toolbox every time but there's nothing we
can do about that.